new fractional index works
